pyproject.toml
[project]
name = "biopyassistant"
version = "0.1.0"
description = "BioPyAssistant is a teaching assistant for Patrick Fuchs and Pierre Poulain's Python programming course for biologists."
readme = "README.md"
authors = [
    { name = "Essmay Touami", email = "essmay.touami@etu.u-paris.fr" },
    { name = "Pierre Poulain", email = "pierre.poulain@cupnet.net" },
]
requires-python = "==3.13.*"
dependencies = [
    "python-dotenv",
    "chromadb",
    "openai",
    "tiktoken",
    "langchain",
    "langchain-openai",
    "langchain-chroma",
    "langchain-core>=1.2.11",
    "langchain-community",
    "langchain-text-splitters",
    "loguru",
    "streamlit",
    "click>=8.3.0",
    "htbuilder>=0.9.0",
    "pydantic-settings>=2.12.0",
    "pyperclip>=1.11.0",
    "pillow>=12.1.1",
    "protobuf>=6.33.5",
    "langchain-openrouter>=0.0.2",
]

[dependency-groups]
dev = [
    "bandit>=1.9.3",
    "jupyterlab>=4.4.9",
    "matplotlib>=3.10.1",
    "plotly>=6.0.0",
    "pre-commit>=4.5.1",
    "prek>=0.3.2",
    "pysankeybeta>=1.4.2",
    "pytest>=9.0.2",
    "ruff>=0.15.1",
    "upsetplot>=0.9.0",
    "watermark>=2.5.0",
]

[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
biopyassistant = ["prompt_templates/*.txt"]

[project.scripts]
parse-clean-markdown = "biopyassistant.core.parse_clean_markdown:main"
create-database = "biopyassistant.core.create_database:main"
query-chatbot = "biopyassistant.core.query_chatbot:main"
